+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Typo in top padding for title
 - Nextclade web component version bump to 1.0.2
+### Added
+- Depth coverage graph component.
 
 ## [v0.6.2]
 ### Fixed